+++ NetherGate/src/com/elmakers/mine/bukkit/plugins/nether/PlayerList.java
@@ -1,9 +1,9 @@
 package com.elmakers.mine.bukkit.plugins.nether;
 
-import com.elmakers.mine.bukkit.plugins.nether.dao.TeleportingPlayer;
+import com.elmakers.mine.bukkit.plugins.nether.dao.NetherPlayer;
 import java.util.ArrayList;
 
-public class PlayerList extends ArrayList<TeleportingPlayer>
+public class PlayerList extends ArrayList<NetherPlayer>
 {
 	public static final long serialVersionUID = 1;